backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0
celery==5.3.4
redis==5.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
geoalchemy2==0.14.2
scikit-learn==1.3.2
numpy==1.26.0
scipy==1.11.3
pandas==2.1.1
gdal==3.7.2
fiona==1.9.5
shapely==2.0.2
rasterio==1.3.9
pyproj==3.6.1
matplotlib==3.8.0
plotly==5.18.0
python-dotenv==1.0.0
pytest==7.4.3
boto3==1.29.0
werkzeug==2.3.7
pyjwt==2.8.0